scrapers/fragment/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
httpx==0.28.1
beautifulsoup4==4.13.5
apscheduler==3.10.4